removed boards and added all_users view to buza app
