--- src/project/boards/templates/boards/answer_view.html
@@ -1,23 +1,0 @@
-<div id="question_view" style="width: 100%;
-    text-align: center; left:50%; padding-left: 1%;background: azure;">
-    <div id="votes" style="width: 5%;float: left;padding-right: 5%;">
-                {% include "boards/answer_votes.html" with parent=answer%}
-    </div>
-<p>answer: {{answer.answer}}</p>
-<p>answer media : {{answer.media}}</p>
-<p>answered by: {{user}}</p>
-<p>answered on: {{answer.created_at}}</p>
-<p>user profile name: {{user.user_profile}}</p>
-<br>
-	{# check if I am the one who asked this question #}
- {% if user.is_authenticated and user == user %}
-        <a href="" > <i class="material-icons" style="color: lightCoral;float:right; padding-left: 5%;">delete</i></a>
-        <a href="" style="float:right; padding-left: 5%;">
-        	<i class="material-icons">edit</i> </a>
- {% endif%}
- {% if user.is_authenticated %}
- <p style=" color: grey; float: center;font-style: italic;">Add Comment</p>
- {% else %}
-	<p><a href="{% url "login" %}" ></a> Log-in</p>  to comment</p>
-{%endif%}
-</div>
