--- src/project/boards/templates/boards/answer_votes.html
@@ -1,44 +1,0 @@
-{% load vote_tags %}
-
-{%if not user.is_authenticated%}
- <div style="padding-top: 55%;">
-  <a href= "login">
-         <i class="material-icons icon" style="color: lightgreen;font-size: 30px;">
-          thumb_up<span class="tooltiptext">Login to vote for this answer</span>
-    </i>
-       </a>
-
-      <div style="color: lightgreen;font-size: 30px;">
-     {% vote_count parent %}
-   </div>
-
-    <a href= "login">
-        <i class="material-icons icon" style="color: yellow;font-size: 30px;">bookmark_outline</i>
-        <span class="tooltiptext" style="background:yellow;">Login vote for this answer</span>
-        </a>
-       </div>
-{% elif  not parent.user == user %}
-	{# user can vote #}
-  <div>
-    <form action="" method="post" class="tooltip">
-      <span class="tooltiptext">This answer was helpful</span>
-    {% csrf_token %}
-      <button type="submit"
-        class="material-icons icon" value="{{answer.id}}" name= 'vote-up-answer' style="color: lightgreen;">
-         thumb_up</button>
-
-       </form>
-
-      <div style="color: lightgreen;font-size: 30px;">
-     {% vote_count parent %}
-   </div>
-
-  <form action="" method="post" class="tooltip">
-         <span class="tooltiptext" style="background: lightBlue;">This answer was not helpful</span>
-         {% csrf_token %}
-          <button type="submit"
-        class="material-icons icon" value="{{answer.id}}" name= 'vote-down-answer' style="color: lightBlue;">
-         thumb_down </button>
-       </form>
-      </div>
-{% endif %}
