--- src/project/boards/templates/boards/filtered_names.html
@@ -1,19 +1,0 @@
-{% load static %}
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-     <link rel='stylesheet' href="{% static 'css/boards.css' %}">
-    <title>{{subject.title}} Questions</title>
-  </head>
-  <body>
-    <h1>{{subject.title}}</h1>
-    <div class="questions">
-    {% for question in questions %}
-    <div class="question">
-      {{ question.title }}
-      </div>
-    {% endfor %}
-    </div>
-  </body>
-</html>
