--- src/project/boards/templates/boards/question_not_found.html
@@ -1,1 +1,0 @@
-<dir>QUESTION NOT FOUND</dir>
