--- src/project/boards/templates/boards/question_votes.html
@@ -1,54 +1,0 @@
-{% load vote_tags %}
-
-{%if not user.is_authenticated%}
- <div>
-  <a href= "login">
-         <i class="material-icons icon" style="color: lightgreen;font-size: 40px;">
-          thumb_up<span class="tooltiptext">Login to vote for this question</span>
-    </i>
-       </a>
-
-      <div style="color: lightgreen;font-size: 40px;">
-     {% vote_count parent %}
-   </div>
-
-    <a href= "login">
-        <i class="material-icons icon" style="color: yellow;font-size: 40px;">bookmark_outline</i>
-        <span class="tooltiptext" style="background: yellow;">Login to bookmark this question</span>
-        </a>
-       </div>
-{% elif  parent.user == user %}
-<div >
-         <form action="" method="post" enctype="multipart/form-data" class="tooltip">
-        <i class="material-icons icon" style="color: yellow;font-size: 40px;">bookmark_outline<input type="submit" value="" name= 'star' ></i>
-        <span class="tooltiptext" style="background: yellow;">Click to mark as favourite question</span>
-        </form>
-       </div>
-
-{% else %}
-	{# user can vote #}
-  <div style="display:inline-block;width: 100%">
-  <form action="" method="post" enctype="multipart/form-data" class="tooltip">
-    <span class="tooltiptext" style="background: lightgreen;">This answer was helpful</span>
-    {% csrf_token %}
-            <button type="submit" class="material-icons icon" name='vote-up-question'
-            style="color: lightgreen;" >
-         thumb_up </button>
-       </form>
-      <a style="color: lightgreen;font-size: 20px;">
-     {% vote_count parent %}</a>
-   <form action="" method="post" enctype="multipart/form-data" class="tooltip">
-        <span class="tooltiptext" style="background: yellow;">Click to mark as favourite question</span>
-        {% csrf_token %}
-        <button type="submit"
-        class="material-icons icon" value="" name= 'star' style="color: yellow; font-size: 20px;">
-        {% is_starred parent user as star%}
-         {% if star %}
-         bookmark
-         {% else %}
-         bookmark_outline
-         {% endif %}
-       </button>
-        </form>
-       </div>
-{% endif %}
