--- src/project/boards/templates/boards/questions.html
@@ -1,46 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% block content %}
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <title>Questions</title>
-  </head>
-  <body>
-    <h1>Questions</h1>
-
-    <div name="all_questions">
-    {% for question in questions %}
-    <div class="question">
-      <div class=user-format>
-        <button type="submit" class="material-icons icon" name='view_user'>account_circle </button>
-          {{ user.username }}
-       <a class="date"> {{question.created_at}}</a>
- </div>
-  <br>
-         <a href="{% url 'view_question' question.board.slug question.id question.slug %}">
-          <br>
-      <div class="question-title">{{ question.title }}
-
-       <i class="material-icons" style="float: right;">share</i>
-  </div>
-      {{question.description}}
-<br>
-
-{% if question.media %}
-media files: <img src="{{question.media.url }}" alt="unable to load image" width="500px">
-<br>
- </a>
-{% endif %}
-board: {{question.board.title}}
-<br>
-       {% block like %}
-                {% include "boards/question_votes.html" with parent=question%}
-         {% endblock %}
-      </div>
-        {% endfor %}
-    </div>
-  </body>
-</html>
-{% endblock %}
