--- src/project/settings/__init__.py
@@ -1,1 +1,0 @@
-from .development import *  # noqa:  F401
