--- src/project/settings/development.py
@@ -1,4 +1,0 @@
-from .base import *  # noqa:  F401
-
-
-DEBUG = True
