--- src/project/settings/production.py
@@ -1,5 +1,0 @@
-from .base import *  # noqa: F401
-
-
-DEBUG = False
-ALLOWED_HOSTS = ['buza.com']
